cpt-lib: repository and source revamp. See log.
Many additions and changes here,
- added support for Mercurial sources
- added support for Mercurial repositories
- adapted git source changes from kiss
- Changed rsync repository behaviour:
A new file has been added to rsync repositories: .rsync_root

This file points to the repository root. If the directory that .rsync_root
points also has an '.rsync' file, this will be accepted as the root directory
and the rsync operation will take place on the root. This ensures that we can
have repository exchanges on a single rsync call, while still allowing for
partial repositories.