pkg_owner(): add function to retrieve the owning package of a file
This function is almost identical to the function with the same name on
kiss. However, instead of saving the owner on a global variable, the function
prints the owning package to the standard output.

The function can be called similar to grep. The usage is:

    pkg_owner GREP_FLAGS FILE_TO_CHECK MANIFEST_LOCATIONS...

If no manifest is specified, the function will use every installed package
on the system database.